+++ src/core/library/library_struct.rs
@@ -69,7 +69,7 @@ impl Library {
     }
 
     /// Filters songs in the library
-    pub fn filter(&self, filter: Filter) -> AlcVec<SongId> {
+    pub fn filter(&self, filter: &Filter) -> AlcVec<SongId> {
         let mut buf = String::new();
         (0..self.songs().len())
             .map(SongId)