+++ src/core/messenger/data.rs
@@ -1,9 +1,5 @@
 use serde::{Deserialize, Serialize};
 
-use raplay::Timestamp;
-
-use crate::core::library::Song;
-
 //===========================================================================//
 //                                   Public                                  //
 //===========================================================================//
@@ -36,30 +32,6 @@ pub enum ErrorKind {
     ExpectedRequestOrControl,
     /// Error occured when trying to do what was requested.
     InternalError,
-}
-
-/// Request someting from the other side.
-#[derive(Debug, Serialize, Deserialize)]
-pub enum Request {
-    /// Request the current playback info.
-    Info,
-}
-
-/// Info about the playback.
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Info {
-    /// Uamp version.
-    pub version: String,
-    /// Song that is now playing.
-    pub now_playing: Option<Song>,
-    /// Length of the playlisg.
-    pub playlist_len: usize,
-    /// Current position in the playlist.
-    pub playlist_pos: Option<usize>,
-    /// True if is currently playing.
-    pub is_playing: bool,
-    /// The timestamp of the current playback.
-    pub timestamp: Option<Timestamp>,
 }
 
 impl Error {