+++ src/core/messenger/mod.rs
@@ -1,4 +1,5 @@
 mod data;
+mod info_msg;
 mod messenger_struct;
 mod msg_message;
 
@@ -6,4 +7,4 @@ mod msg_message;
 //                                   Public                                  //
 //===========================================================================//
 
-pub use self::{data::*, messenger_struct::*, msg_message::*};
+pub use self::{data::*, info_msg::*, messenger_struct::*, msg_message::*};