+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add new filtering options.
 - Add new options to edit the current playlist (`queue`, `play-next`,
   `push-with-cur`, `flatten`)
+- Add option to list songs (`list`)
 - Add 1 to playlist position when displaying with `-Info` (so it starts from 1
   instead of 0)
 - Log playback errors.